+++ src/main/java/talkPick/domain/admin/application/command/TopicCommandService.java
@@ -1,11 +1,10 @@
-package talkPick.domain.topic.application;
+package talkPick.domain.admin.application.command;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import talkPick.domain.topic.port.out.TopicLikeHistoryCommandRepositoryPort;
 import talkPick.domain.topic.port.out.TopicStatQueryRepositoryPort;
-import talkPick.global.error.ErrorCode;
 import talkPick.domain.topic.port.in.TopicCommandUseCase;
 
 @Service
