--- src/main/java/talkPick/domain/topic/application/admin/AdminTopicCommandService.java
@@ -1,65 +1,0 @@
-package talkPick.domain.topic.application.admin;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import talkPick.domain.topic.adapter.in.mapper.TopicReqMapper;
-import talkPick.domain.topic.adapter.in.dto.TopicReqDTO;
-import talkPick.domain.topic.adapter.out.dto.TopicResDTO;
-import talkPick.domain.admin.domain.Admin;
-import talkPick.domain.topic.domain.Topic;
-import talkPick.domain.topic.port.in.admin.AdminTopicCommandUseCase;
-import talkPick.domain.admin.port.out.AdminQueryRepositoryPort;
-import talkPick.domain.topic.port.out.TopicStatCommandRepositoryPort;
-import talkPick.domain.topic.port.out.TopicQueryRepositoryPort;
-import talkPick.domain.topic.port.out.admin.AdminTopicCommandRepositoryPort;
-
-@Service
-@Transactional
-@RequiredArgsConstructor
-public class AdminTopicCommandService implements AdminTopicCommandUseCase {
-    private final AdminTopicCommandRepositoryPort adminTopicCommandRepositoryPort;
-    private final TopicStatCommandRepositoryPort topicStatCommandRepositoryPort;
-    private final TopicQueryRepositoryPort topicQueryRepositoryPort;
-    private final AdminQueryRepositoryPort adminQueryRepositoryPort;
-    private final TopicReqMapper topicReqMapper;
-
-    @Override
-    public TopicResDTO.Topic createTopicByAdmin(Long adminId, TopicReqDTO.CreateTopic createTopic) {
-        Topic topic = adminTopicCommandRepositoryPort.save(
-                Topic.create( // topic 생성
-                        createTopic.title(),
-                        createTopic.detail(),
-                        createTopic.thumbnail(),
-                        createTopic.icon(),
-                        getValidAdmin(adminId)
-                )
-        );
-        topicStatCommandRepositoryPort.save(topic.getId()); // topic 통계 생성
-        return topicReqMapper.toTopicResponse(topic);
-    }
-
-    @Override
-    public TopicResDTO.Topic updateTopicByAdmin(Long topicId, TopicReqDTO.CreateTopic updateTopic) {
-
-        // 해당 topic 찾아 업데이트
-        Topic topic = getValidTopic(topicId);
-        topic.update(updateTopic.title(), updateTopic.detail(), updateTopic.thumbnail(), updateTopic.icon());
-
-        Topic updated = adminTopicCommandRepositoryPort.save(topic);
-        return topicReqMapper.toTopicResponse(updated);
-    }
-
-    @Override
-    public void deleteTopicByAdmin(Long topicId) {
-        adminTopicCommandRepositoryPort.delete(topicId);
-    }
-
-    private Admin getValidAdmin(Long adminId) {
-        return adminQueryRepositoryPort.findAdminById(adminId);
-    }
-
-    private Topic getValidTopic(Long topicId) {
-        return topicQueryRepositoryPort.findTopicById(topicId);
-    }
-}
